requirements.txt
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
openpyxl==3.1.5
scikit-learn==1.7.2
joblib==1.5.2
matplotlib==3.10.3
streamlit==1.50.0
shap==0.42.1
plotly==5.20.0
python-dotenv==1.0.0
hopsworks==4.4.2
xgboost==2.1.2


